+++ lib/generators/rails_url_shortener/templates/initializer.rb
@@ -15,3 +15,5 @@ RailsUrlShortener.save_bots_visits = false               # if save bots visits
 RailsUrlShortener.save_visits = true                     # if save visits
 RailsUrlShortener.disable_url_key_updates = false        # if true, prevents the key from being updated after creation
 RailsUrlShortener.forward_query_params = false           # if true, forwards query params from short URL to redirect destination
+RailsUrlShortener.ip_lookup_backend = RailsUrlShortener::IpLookup::IP_API_COM
+# RailsUrlShortener.ip_lookup_api_key = ENV['IP_LOOKUP_API_KEY']